bopp/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
astroid==3.3.4
attrs==24.2.0
autopep8==2.3.1
boto3==1.35.20
botocore==1.35.20
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
dill==0.3.9
Django==5.0
django-cors-headers==4.4.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-ratelimit==4.1.0
django-sns-mobile-push-notification==0.9
django-storages==1.14.4
django-twilio==0.14.3.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
gunicorn==23.0.0
httplib2==0.22.0
idna==3.7
isort==5.13.2
jmespath==1.0.1
jwcrypto==1.5.6
mccabe==0.7.0
msgpack==1.0.8
multidict==6.1.0
node==1.2.2
oauth2client==4.1.3
odict==1.9.0
packaging==24.1
phonenumbers==8.13.45
pillow==10.4.0
platformdirs==4.3.6
plumber==1.7
proto-plus==1.24.0
protobuf==4.25.4
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pylint==3.3.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-jwt==4.1.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
s3transfer==0.10.2
setuptools==71.1.0
six==1.16.0
sqlparse==0.5.1
tomlkit==0.13.2
twilio==9.3.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
yarl==1.11.1
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==6.4.post2
zope.lifecycleevent==5.0
zope.proxy==5.2
